Rename rebuilt modal to Modal.rebuilt
