+++ packages/components/src/Modal/index.ts
@@ -1,3 +1,3 @@
-export { useModalContext } from "./ModalContext";
-export { Modal } from "./Modal";
+export { useModalContext } from "./ModalContext.rebuilt";
+export { Modal } from "./Modal.rebuilt";
 export { useModalStyles } from "./useModalStyles";
